Split benchamrks
